Display owns the HWND
Display owns the HWND so that the native display (DC) belongs to the
Window.  Otherwise webkit would draw directly to the Windows desktop.

Also, create a thread to create the window and service the message
pump for that Window.  Tear-down is not addressed yet.
